requirements.txt
trame
trame-vuetify
trame-vtk
vtk-module